feat(6780): extract docs contract wave4 harness
