+++ crates/kamn-core/tests/docs_contract_wave4_harness/key_lifecycle_audit_trails_docs.rs
@@ -1,0 +1,58 @@
+const DOC: &str = include_str!("../../../../docs/foundation/key-lifecycle-audit-trails.md");
+
+#[test]
+fn doc_contains_audit_record_schema_and_verification_rules() {
+    assert!(DOC.contains("## Scope Delivered"));
+    assert!(DOC.contains("KeyLifecycleAuditRecord"));
+    assert!(DOC.contains("KeyLifecycle::audit_records()"));
+    assert!(DOC.contains("KeyLifecycle::verify_audit_records(...)"));
+    assert!(DOC.contains("KeyLifecycleAuditError"));
+}
+
+#[test]
+fn doc_contains_tamper_evident_chain_requirements() {
+    assert!(DOC.contains("## Tamper-Evident Rules"));
+    assert!(DOC.contains("genesis marker `GENESIS`"));
+    assert!(DOC.contains("Sequence IDs must be contiguous and start at `1`."));
+    assert!(DOC.contains("Verification fails when sequence continuity, chain links, or record hashes are inconsistent."));
+}
+
+#[test]
+fn regression_requires_chain_link_mismatch_detection_rule() {
+    // Regression: #158
+    assert!(DOC.contains("chain links"));
+}
+
+#[test]
+fn doc_contains_lifecycle_operator_binding_audit_evidence_contract() {
+    assert!(DOC.contains("## DID Lifecycle Operator-Binding Audit Evidence Contract (Issue #890)"));
+    assert!(DOC.contains("lifecycle_operator_binding_contract.py"));
+    assert!(DOC.contains("generate_lifecycle_operator_binding_evidence_bundle.sh"));
+    assert!(DOC.contains("check_lifecycle_operator_binding_policy.sh"));
+    assert!(DOC.contains("run_lifecycle_operator_binding_contract_lane.sh"));
+    assert!(DOC.contains("did_lifecycle_operator_binding_reason_codes:GO:v1"));
+}
+
+#[test]
+fn regression_doc_marks_missing_keys_or_decision_drift_fail_closed_policy() {
+    // Regression: #890
+    assert!(DOC.contains("Regression: #890"));
+}
+
+#[test]
+fn doc_contains_secure_provider_key_lifecycle_evidence_contract() {
+    assert!(
+        DOC.contains("## Secure-Provider Key Rotation/Revocation Evidence Contract (Issue #988)")
+    );
+    assert!(DOC.contains("secure_provider_key_lifecycle_contract.py"));
+    assert!(DOC.contains("generate_secure_provider_key_lifecycle_evidence_bundle.sh"));
+    assert!(DOC.contains("check_secure_provider_key_lifecycle_policy.sh"));
+    assert!(DOC.contains("run_secure_provider_key_lifecycle_contract_lane.sh"));
+    assert!(DOC.contains("secure_provider_key_lifecycle_reason_codes:GO:v1"));
+}
+
+#[test]
+fn regression_doc_marks_secure_provider_lifecycle_fail_closed_policy() {
+    // Regression: #988
+    assert!(DOC.contains("Regression: #988"));
+}
