+++ crates/kamn-core/tests/docs_contract_wave4_harness/task_swarm_dag_docs.rs
@@ -1,0 +1,61 @@
+const OPERATIONS_DOC: &str = include_str!("../../../../docs/foundation/task-operations.md");
+const STATE_MACHINE_DOC: &str = include_str!("../../../../docs/foundation/task-state-machine.md");
+
+#[test]
+fn docs_define_swarm_dag_command_surface() {
+    assert!(OPERATIONS_DOC.contains("SwarmTaskDraft"));
+    assert!(OPERATIONS_DOC.contains("submit_swarm_tasks(drafts)"));
+    assert!(OPERATIONS_DOC.contains("ready_tasks()"));
+    assert!(OPERATIONS_DOC.contains("DependencyNotSatisfied"));
+}
+
+#[test]
+fn docs_define_dependency_aware_transition_gates() {
+    assert!(STATE_MACHINE_DOC.contains("## Dependency-Aware Transition Gates"));
+    assert!(STATE_MACHINE_DOC.contains("TaskOperationEngine::start_work"));
+    assert!(STATE_MACHINE_DOC
+        .contains("all declared dependencies must already be in `Completed` state."));
+    assert!(STATE_MACHINE_DOC.contains("`InputRequired`"));
+    assert!(STATE_MACHINE_DOC.contains("InputRequired -> InProgress | Failed | Cancelled"));
+}
+
+#[test]
+fn regression_requires_cyclic_and_premature_transition_guards() {
+    // Regression: #472
+    assert!(OPERATIONS_DOC.contains("Regression: #472"));
+    assert!(STATE_MACHINE_DOC.contains("Regression: #472"));
+}
+
+#[test]
+fn docs_define_bounded_graph_benchmark_lane() {
+    assert!(OPERATIONS_DOC.contains("bounded graph benchmark"));
+    assert!(OPERATIONS_DOC.contains("cargo test -p kamn-core --test swarm_task_dag"));
+}
+
+#[test]
+fn docs_define_snapshot_recovery_validation_rules() {
+    assert!(OPERATIONS_DOC.contains("export_snapshot()"));
+    assert!(OPERATIONS_DOC.contains("restore_snapshot(snapshot)"));
+    assert!(OPERATIONS_DOC.contains("schema version mismatch is rejected."));
+    assert!(STATE_MACHINE_DOC.contains("Snapshot restore invariants"));
+}
+
+#[test]
+fn regression_requires_tampered_snapshot_rejection_rule() {
+    // Regression: #502
+    assert!(OPERATIONS_DOC.contains("Regression: #502"));
+    assert!(STATE_MACHINE_DOC.contains("Regression: #502"));
+}
+
+#[test]
+fn docs_define_snapshot_roundtrip_benchmark_lane() {
+    assert!(OPERATIONS_DOC.contains("snapshot roundtrip benchmark"));
+    assert!(OPERATIONS_DOC.contains("cargo test -p kamn-core --test task_operation_snapshot"));
+}
+
+#[test]
+fn regression_requires_input_required_operation_surface() {
+    // Regression: #573
+    assert!(OPERATIONS_DOC.contains("request_input(task_id, actor, reason)"));
+    assert!(OPERATIONS_DOC.contains("emits `InputRequired` notice."));
+}
